refactor(backend): extract shared EXIF/GPS parsing into app/core/exif.py
Deduplicates EXIF/GPS extraction between image_processor.py and
vips_processor.py into a shared module, and replaces print() calls
with logging in both processors.
