+++ backend/app/core/image_processor.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import asyncio
 import contextlib
+import logging
 import os
 import typing
 import uuid
@@ -14,7 +15,9 @@ from PIL import Image
 from PIL.ExifTags import TAGS
 
 from app.config import settings
-from app.core.location_service import location_service
+from app.core.exif import extract_comprehensive_exif
+
+logger = logging.getLogger(__name__)
 
 
 class ImageProcessor:
@@ -29,7 +32,7 @@ class ImageProcessor:
             exif_data: dict[str, Any] = {"width": img.width, "height": img.height}
             try:
                 exif_dict = piexif.load(image_path)
-                comprehensive_data = await self._extract_comprehensive_exif(exif_dict)
+                comprehensive_data = await extract_comprehensive_exif(exif_dict)
                 exif_data.update(comprehensive_data)
             except Exception:
                 if hasattr(img, "_getexif") and img._getexif() is not None:
@@ -41,96 +44,9 @@ class ImageProcessor:
         """Extract comprehensive EXIF data from image including timezone and GPS."""
         try:
             return await self._read_exif_from_image(image_path)
-        except Exception as e:
-            print(f"Error extracting EXIF data: {e}")
+        except Exception:
+            logger.exception("Error extracting EXIF data")
             return {}
-
-    async def _extract_comprehensive_exif(self, exif_dict: dict) -> dict[str, Any]:
-        """Extract comprehensive EXIF data using piexif."""
-        data: dict[str, Any] = {}
-
-        # Extract from 0th IFD (main image)
-        if "0th" in exif_dict:
-            ifd = exif_dict["0th"]
-
-            # Camera make and model
-            if piexif.ImageIFD.Make in ifd:
-                data["camera_make"] = ifd[piexif.ImageIFD.Make].decode().strip()
-            if piexif.ImageIFD.Model in ifd:
-                data["camera_model"] = ifd[piexif.ImageIFD.Model].decode().strip()
-
-            # Date and time
-            if piexif.ImageIFD.DateTime in ifd:
-                try:
-                    dt_str = ifd[piexif.ImageIFD.DateTime].decode()
-                    data["date_taken"] = datetime.strptime(dt_str, "%Y:%m:%d %H:%M:%S")
-                except (ValueError, AttributeError):
-                    pass
-
-        # Extract from Exif IFD (detailed camera settings)
-        if "Exif" in exif_dict:
-            ifd = exif_dict["Exif"]
-
-            # Original date (often more accurate)
-            if piexif.ExifIFD.DateTimeOriginal in ifd:
-                try:
-                    dt_str = ifd[piexif.ExifIFD.DateTimeOriginal].decode()
-                    data["date_taken"] = datetime.strptime(dt_str, "%Y:%m:%d %H:%M:%S")
-                except (ValueError, AttributeError):
-                    pass
-
-            # Camera settings
-            if piexif.ExifIFD.ISOSpeedRatings in ifd:
-                data["iso"] = ifd[piexif.ExifIFD.ISOSpeedRatings]
-
-            if piexif.ExifIFD.FNumber in ifd:
-                f_num = ifd[piexif.ExifIFD.FNumber]
-                if isinstance(f_num, tuple) and len(f_num) == 2 and f_num[1] != 0:
-                    data["aperture"] = float(f_num[0] / f_num[1])
-
-            if piexif.ExifIFD.ExposureTime in ifd:
-                exp_time = ifd[piexif.ExifIFD.ExposureTime]
-                if isinstance(exp_time, tuple) and len(exp_time) == 2:
-                    if exp_time[0] == 1:
-                        data["shutter_speed"] = f"1/{exp_time[1]}"
-                    else:
-                        data["shutter_speed"] = f"{exp_time[0]}/{exp_time[1]}"
-
-            if piexif.ExifIFD.FocalLength in ifd:
-                focal = ifd[piexif.ExifIFD.FocalLength]
-                if isinstance(focal, tuple) and len(focal) == 2 and focal[1] != 0:
-                    data["focal_length"] = int(focal[0] / focal[1])
-
-            # Lens information
-            if piexif.ExifIFD.LensModel in ifd:
-                data["lens"] = ifd[piexif.ExifIFD.LensModel].decode().strip()
-
-            # Timezone information from SubSecTimeOriginal or OffsetTime
-            timezone_offset = None
-            if piexif.ExifIFD.OffsetTimeOriginal in ifd:
-                try:
-                    timezone_offset = (
-                        ifd[piexif.ExifIFD.OffsetTimeOriginal].decode().strip()
-                    )
-                    data["timezone"] = timezone_offset
-                except (ValueError, AttributeError):
-                    pass
-
-        # Extract GPS data
-        if "GPS" in exif_dict:
-            gps_data = self._extract_enhanced_gps_data(exif_dict["GPS"])
-            data.update(gps_data)
-
-            # Try to get location name if we have coordinates
-            if "location_lat" in gps_data and "location_lon" in gps_data:
-                location_info = await location_service.reverse_geocode(
-                    gps_data["location_lat"], gps_data["location_lon"]
-                )
-                if location_info:
-                    data["location_name"] = location_info.location_name
-                    data["location_address"] = location_info.location_address
-
-        return data
 
     def _extract_basic_exif(self, exif: dict[typing.Any, typing.Any]) -> dict[str, Any]:
         """Fallback basic EXIF extraction using PIL."""
@@ -181,73 +97,6 @@ class ImageProcessor:
         return data
 
     @staticmethod
-    def _dms_to_decimal(
-        dms_tuple: tuple[tuple[int, int], tuple[int, int], tuple[int, int]] | None,
-    ) -> float | None:
-        if not dms_tuple or len(dms_tuple) != 3:
-            return None
-        degrees = (
-            float(dms_tuple[0][0] / dms_tuple[0][1]) if dms_tuple[0][1] != 0 else 0
-        )
-        minutes = (
-            float(dms_tuple[1][0] / dms_tuple[1][1]) if dms_tuple[1][1] != 0 else 0
-        )
-        seconds = (
-            float(dms_tuple[2][0] / dms_tuple[2][1]) if dms_tuple[2][1] != 0 else 0
-        )
-        return degrees + (minutes / 60.0) + (seconds / 3600.0)
-
-    def _parse_enhanced_gps(
-        self, gps_info: dict[int | str, typing.Any]
-    ) -> dict[str, typing.Any]:
-        gps_data: dict[str, typing.Any] = {}
-
-        if (
-            piexif.GPSIFD.GPSLatitude in gps_info
-            and piexif.GPSIFD.GPSLatitudeRef in gps_info
-        ):
-            lat_decimal = self._dms_to_decimal(gps_info[piexif.GPSIFD.GPSLatitude])
-            if lat_decimal is not None:
-                lat_ref = gps_info[piexif.GPSIFD.GPSLatitudeRef].decode()
-                gps_data["location_lat"] = (
-                    -lat_decimal if lat_ref in ("S", "s") else lat_decimal
-                )
-
-        if (
-            piexif.GPSIFD.GPSLongitude in gps_info
-            and piexif.GPSIFD.GPSLongitudeRef in gps_info
-        ):
-            lon_decimal = self._dms_to_decimal(gps_info[piexif.GPSIFD.GPSLongitude])
-            if lon_decimal is not None:
-                lon_ref = gps_info[piexif.GPSIFD.GPSLongitudeRef].decode()
-                gps_data["location_lon"] = (
-                    -lon_decimal if lon_ref in ("W", "w") else lon_decimal
-                )
-
-        if piexif.GPSIFD.GPSAltitude in gps_info:
-            alt_tuple = gps_info[piexif.GPSIFD.GPSAltitude]
-            if alt_tuple and len(alt_tuple) == 2 and alt_tuple[1] != 0:
-                altitude = float(alt_tuple[0] / alt_tuple[1])
-                if (
-                    piexif.GPSIFD.GPSAltitudeRef in gps_info
-                    and gps_info[piexif.GPSIFD.GPSAltitudeRef] == 1
-                ):
-                    altitude = -altitude
-                gps_data["altitude"] = altitude
-
-        return gps_data
-
-    def _extract_enhanced_gps_data(
-        self, gps_info: dict[int | str, typing.Any]
-    ) -> dict[str, typing.Any]:
-        """Extract enhanced GPS coordinates and altitude from EXIF GPS info using piexif."""
-        try:
-            return self._parse_enhanced_gps(gps_info)
-        except Exception as e:
-            print(f"Error extracting enhanced GPS data: {e}")
-            return {}
-
-    @staticmethod
     def _dms_to_decimal_basic(dms: tuple[float, float, float], ref: str) -> float:
         decimal = float(dms[0]) + float(dms[1]) / 60.0 + float(dms[2]) / 3600.0
         return -decimal if ref in ("S", "W") else decimal
@@ -266,8 +115,8 @@ class ImageProcessor:
                     "location_lat": self._dms_to_decimal_basic(lat_dms, lat_ref),
                     "location_lon": self._dms_to_decimal_basic(lon_dms, lon_ref),
                 }
-        except Exception as e:
-            print(f"Error extracting GPS data: {e}")
+        except Exception:
+            logger.exception("Error extracting GPS data")
         return {}
 
     async def process_image(
@@ -342,8 +191,8 @@ class ImageProcessor:
         try:
             with Image.open(original_path) as opened_img:
                 return self._build_variants(opened_img, file_id)
-        except Exception as e:
-            print(f"Error generating responsive variants: {e}")
+        except Exception:
+            logger.exception("Error generating responsive variants")
             raise
 
     @staticmethod
@@ -365,8 +214,8 @@ class ImageProcessor:
                     orientation = exif.get(0x0112)
                     if orientation:
                         return self._apply_orientation(img, orientation)
-        except Exception as e:
-            print(f"Error auto-rotating image: {e}")
+        except Exception:
+            logger.exception("Error auto-rotating image")
         return img
 
     def _resize_image(self, img: Image.Image, target_size: int) -> Image.Image:
@@ -402,8 +251,8 @@ class ImageProcessor:
             if file_path and os.path.exists(file_path):
                 try:
                     os.remove(file_path)
-                except Exception as e:
-                    print(f"Error deleting file {file_path}: {e}")
+                except Exception:
+                    logger.exception("Error deleting file %s", file_path)
 
     @staticmethod
     def get_image_url(
